Gate some crates behind features in mas-http
