+++ crates/http/Cargo.toml
@@ -6,14 +6,14 @@ edition = "2021"
 license = "Apache-2.0"
 
 [dependencies]
-axum = "0.5.13"
+axum = { version = "0.5.13", optional = true }
 bytes = "1.2.1"
 futures-util = "0.3.21"
 headers = "0.3.7"
 http = "0.2.8"
 http-body = "0.4.5"
 hyper = "0.14.20"
-hyper-rustls = { version = "0.23.0", features = ["http1", "http2", "rustls-native-certs"], default-features = false }
+hyper-rustls = { version = "0.23.0", features = ["http1", "http2", "rustls-native-certs"], default-features = false, optional = true }
 once_cell = "1.13.0"
 opentelemetry = "0.17.0"
 opentelemetry-http = "0.6.0"
@@ -23,14 +23,19 @@ serde = "1.0.142"
 serde_json = "1.0.83"
 serde_urlencoded = "0.7.1"
 thiserror = "1.0.32"
-tokio = { version = "1.20.1", features = ["sync", "parking_lot"] }
+tokio = { version = "1.20.1", optional = true }
 tower = { version = "0.4.13", features = ["timeout", "limit"] }
-tower-http = { version = "0.3.4", features = ["follow-redirect", "decompression-full", "set-header", "compression-full", "cors"] }
+tower-http = { version = "0.3.4", features = ["follow-redirect", "decompression-full", "set-header", "compression-full", "cors", "util"] }
 tracing = "0.1.36"
 tracing-opentelemetry = "0.17.4"
 
 [dev-dependencies]
 anyhow = "1.0.62"
 serde = { version = "1.0.142", features = ["derive"] }
-tokio = { version = "1.20.1", features = ["macros"] }
+tokio = { version = "1.20.1", features = ["macros", "rt"] }
 tower = { version = "0.4.13", features = ["util"] }
+
+[features]
+default = []
+axum = ["dep:axum"]
+client = ["dep:hyper-rustls", "hyper/tcp", "tokio", "tokio/sync", "tokio/parking_lot"]
