+++ crates/http/src/layers/otel/on_response.rs
@@ -14,12 +14,10 @@
 
 use headers::{ContentLength, HeaderMapExt};
 use http::Response;
+#[cfg(feature = "client")]
 use hyper::client::connect::HttpInfo;
 use opentelemetry::trace::SpanRef;
-use opentelemetry_semantic_conventions::trace::{
-    HTTP_RESPONSE_CONTENT_LENGTH, HTTP_STATUS_CODE, NET_HOST_IP, NET_HOST_PORT, NET_PEER_IP,
-    NET_PEER_PORT,
-};
+use opentelemetry_semantic_conventions::trace as SC;
 
 pub trait OnResponse<R> {
     fn on_response(&self, span: &SpanRef<'_>, response: &R);
@@ -37,21 +35,22 @@ pub struct OnHttpResponse;
 
 impl<B> OnResponse<Response<B>> for OnHttpResponse {
     fn on_response(&self, span: &SpanRef<'_>, response: &Response<B>) {
-        span.set_attribute(HTTP_STATUS_CODE.i64(i64::from(response.status().as_u16())));
+        span.set_attribute(SC::HTTP_STATUS_CODE.i64(i64::from(response.status().as_u16())));
 
         if let Some(ContentLength(content_length)) = response.headers().typed_get() {
             if let Ok(content_length) = content_length.try_into() {
-                span.set_attribute(HTTP_RESPONSE_CONTENT_LENGTH.i64(content_length));
+                span.set_attribute(SC::HTTP_RESPONSE_CONTENT_LENGTH.i64(content_length));
             }
         }
 
+        #[cfg(feature = "client")]
         // Get local and remote address from hyper's HttpInfo injected by the
         // HttpConnector
         if let Some(info) = response.extensions().get::<HttpInfo>() {
-            span.set_attribute(NET_PEER_IP.string(info.remote_addr().ip().to_string()));
-            span.set_attribute(NET_PEER_PORT.i64(info.remote_addr().port().into()));
-            span.set_attribute(NET_HOST_IP.string(info.local_addr().ip().to_string()));
-            span.set_attribute(NET_HOST_PORT.i64(info.local_addr().port().into()));
+            span.set_attribute(SC::NET_PEER_IP.string(info.remote_addr().ip().to_string()));
+            span.set_attribute(SC::NET_PEER_PORT.i64(info.remote_addr().port().into()));
+            span.set_attribute(SC::NET_HOST_IP.string(info.local_addr().ip().to_string()));
+            span.set_attribute(SC::NET_HOST_PORT.i64(info.local_addr().port().into()));
         }
     }
 }
